Changes connected to Utils refactor.
Refactoring the Utils directory and the utils.ts file resulted in changes to imports. I had to change the paths.
